// src/lib/types.ts
export interface Auth {
  user?: {
    id: string;
    username: string;
    role: string;
    isAdmin: boolean;
  };
  shareToken?: {
    shareType?: number;
    websiteId?: string;
    websiteIds?: string[];
    boardId?: string;
    pixelId?: string;
    pixelIds?: string[];
    linkId?: string;
    linkIds?: string[];
    parameters?: ShareParameters;
  };
}

export type ShareTheme = 'light' | 'dark';

export interface ShareParameters {
  allowFilter?: boolean;
  theme?: ShareTheme;
  [key: string]: boolean | ShareTheme | undefined;
}